fix: #WB-2473, retrieve all the content of the template columns
The previous method only retrieved the content of a column ifit was embedded in a paragraph tag.
However, when the user format the content of a template column, the content is not embedded in a p anymore but in various divs with style.
Now we retrieve all the content o the column (except the title) and pass it to the render method.